+++ flask_sillywalk/sillywalk.py
@@ -114,22 +114,110 @@ class SwaggerApiRegistry(object):
                 if self.models[c.__name__].get("required") is None:
                     self.models[c.__name__]["required"] = []
                 self.models[c.__name__]["required"].append(arg)
-                #self.models[c.__name__]["required"][arg] = {"required": True}
+                # self.models[c.__name__]["required"][arg] = {"required": True}
             for k, v in defaults:
                 self.models[c.__name__]["properties"][k] = {"default": v}
             return c
 
         return inner_func
 
-    def register(
-            self,
+    def _register(self,
+                  path,
+                  f,
+                  method,
+                  content_type,
+                  parameters,
+                  responseMessages,
+                  nickname,
+                  notes,
+                  bp):
+
+        if self.app is None:
+            raise SwaggerRegistryError(
+                "You need to initialize {0} with a Flask app".format(
+                    self.__class__.__name__))
+
+        # register views on blueprints
+        app = self.app
+
+        if bp:
+            app = bp
+
+        app.add_url_rule(
             path,
-            method="GET",
-            content_type="application/json",
-            parameters=[],
-            responseMessages=[],
-            nickname=None,
-            notes=None):
+            f.__name__,
+            f,
+            methods=[method])
+
+        api = Api(
+            method=f,
+            path=path.replace(self.basepath, ""),
+            httpMethod=method,
+            params=parameters,
+            responseMessages=responseMessages,
+            nickname=nickname,
+            notes=notes)
+
+        if api.resource not in self.app.view_functions:
+            for fmt in SUPPORTED_FORMATS:
+                route = "{0}/{1}.{2}".format(self.basepath.rstrip("/"),
+                                             api.resource, fmt)
+                if route not in self.registered_routes:
+                    self.registered_routes.append(route)
+                    self.app.add_url_rule(
+                        route,
+                        api.resource,
+                        self.show_resource(api.resource))
+
+        if self.r[api.resource].get(api.path) is None:
+            self.r[api.resource][api.path] = list()
+        self.r[api.resource][api.path].append(api)
+
+    def add_register(self,
+                     path,
+                     f,
+                     method="GET",
+                     content_type="application/json",
+                     parameters=[],
+                     responseMessages=[],
+                     nickname=None,
+                     notes=None,
+                     bp=None):
+        """
+        Registers an API endpoint.
+
+        Usage:
+
+        >>> def get_cheese(cheesename):
+        >>>     # some function
+        >>> my_registry.add_register(
+        ...     "/api/v1/cheese/<cheeseName>",
+        ...     get_cheese,
+        ...     parameters=[ApiParameter(
+        ...         name="cheeseName",
+        ...         description="The name of the cheese to fetch",
+        ...         required=True,
+        ...         dataType="str",
+        ...         paramType="path",
+        ...         allowMultiple=False)],
+        ...     notes='For getting cheese, you know...',
+        ...     responseMessages=[
+        ...         ApiErrorResponse(400, "Sorry, we're fresh out of that cheese."),
+        ...         ApiErrorResponse(418, "I'm actually a teapot")]))
+
+        """
+        self._register(path, f, method, content_type, parameters,
+                       responseMessages, nickname, notes, bp)
+
+    def register(self,
+                 path,
+                 method="GET",
+                 content_type="application/json",
+                 parameters=[],
+                 responseMessages=[],
+                 nickname=None,
+                 notes=None,
+                 bp=None):
         """
         Registers an API endpoint.
 
@@ -152,43 +240,9 @@ class SwaggerApiRegistry(object):
         >>>     # some function
 
         """
-
         def inner_func(f):
-            if self.app is None:
-                raise SwaggerRegistryError(
-                    "You need to initialize {0} with a Flask app".format(
-                        self.__class__.__name__))
-
-            self.app.add_url_rule(
-                path,
-                f.__name__,
-                f,
-                methods=[method])
-
-            api = Api(
-                method=f,
-                path=path.replace(self.basepath, ""),
-                httpMethod=method,
-                params=parameters,
-                responseMessages=responseMessages,
-                nickname=nickname,
-                notes=notes)
-
-            if api.resource not in self.app.view_functions:
-                for fmt in SUPPORTED_FORMATS:
-                    route = "{0}/{1}.{2}".format(self.basepath.rstrip("/"),
-                                                 api.resource, fmt)
-                    if route not in self.registered_routes:
-                        self.registered_routes.append(route)
-                        self.app.add_url_rule(
-                            route,
-                            api.resource,
-                            self.show_resource(api.resource))
-
-            if self.r[api.resource].get(api.path) is None:
-                self.r[api.resource][api.path] = list()
-            self.r[api.resource][api.path].append(api)
-
+            self._register(path, f, method, content_type, parameters,
+                           responseMessages, nickname, notes, bp)
         return inner_func
 
     def show_resource(self, resource):