+++ flask_sillywalk/tests/test_sillywalk.py
@@ -3,7 +3,7 @@ import os
 import json
 import unittest
 
-from flask import Flask, make_response, request
+from flask import Flask, make_response, request, Blueprint
 from flask.ext.sillywalk import SwaggerApiRegistry, ApiParameter, ApiErrorResponse
 from flask.ext.sillywalk.compat import s
 
@@ -16,6 +16,7 @@ class TestDecorators(unittest.TestCase):
 
     def _create_app(self):
         app = Flask("foobar")
+        app.config['TESTING'] = True
         url = os.environ.get("URL", "localhost:5000")
         registry = SwaggerApiRegistry(app, baseurl="http://{}/api/v1".format(url))
 
@@ -30,6 +31,11 @@ class TestDecorators(unittest.TestCase):
 
         return app, registry
 
+    def _create_blueprint(self, app):
+        bp = Blueprint("blueish", "bluishns")
+        app.register_blueprint(bp)
+        return bp
+
     def _create_model(self, registerModel):
         @registerModel()
         class SomeCrazyClass(object):
@@ -43,7 +49,33 @@ class TestDecorators(unittest.TestCase):
             def say_happy_birthday(self):
                 raise HappyBirthdayException("Chances are it's not your birthday.")
 
-    def _create_register(self, register):
+    def _create_add_register(self, add_register, app, bp=None):
+        def get_cheese(cheeseName):
+            """Gets cheese, just like the name says."""
+            return json.dumps({"response": "OK", "message": "Sorry, we're fresh out of {0}!".format(cheeseName)})
+
+        add_register(
+            "/api/v1/cheese/<cheeseName>",
+            get_cheese,
+            parameters=[
+                ApiParameter(
+                    name="cheeseName",
+                    description="The name of the cheese to fetch",
+                    required=True,
+                    dataType="str",
+                    paramType="path",
+                    allowMultiple=False)],
+            notes='For getting cheese, you know...',
+            responseMessages=[
+                ApiErrorResponse(400, "Sorry, we're fresh out of that cheese."),
+                ApiErrorResponse(418, "I'm actually a teapot")
+            ],
+            bp=bp)
+        # we have register our blueprint again to activate our added route
+        if bp:
+            app.register_blueprint(bp)
+
+    def _create_register(self, register, app, bp=None):
         @register(
             "/api/v1/cheese/<cheeseName>",
             parameters=[
@@ -58,7 +90,8 @@ class TestDecorators(unittest.TestCase):
             responseMessages=[
                 ApiErrorResponse(400, "Sorry, we're fresh out of that cheese."),
                 ApiErrorResponse(418, "I'm actually a teapot")
-            ])
+            ],
+            bp=bp)
         def get_cheese(cheeseName):
             """Gets cheese, just like the name says."""
             return json.dumps({"response": "OK", "message": "Sorry, we're fresh out of {0}!".format(cheeseName)})
@@ -73,7 +106,8 @@ class TestDecorators(unittest.TestCase):
                     required=True,
                     dataType="int",
                     paramType="path",
-                    allowMultiple=False)])
+                    allowMultiple=False)],
+            bp=bp)
         def get_a_holy_hand_grenade(number):
             """Gets one or more holy hand grenades, just like the name says."""
             return json.dumps("Fetching {} holy hand grenades".format(number))
@@ -95,12 +129,17 @@ class TestDecorators(unittest.TestCase):
                     required=False,
                     dataType="str",
                     paramType="query",
-                    allowMultiple=False)])
+                    allowMultiple=False)],
+            bp=bp)
         def toss_the_grenade(number):
             """Toss the holy hand grenade after {number} seconds."""
             target = request.args.get("target", "FOO")
             return json.dumps("Waiting {} seconds to toss the grenade at {}.".format(
                 number, target))
+
+        # we have register our blueprint again to activate our added route
+        if bp:
+            app.register_blueprint(bp)
 
     def test_model(self):
         app, registry = self._create_app()
@@ -119,7 +158,7 @@ class TestDecorators(unittest.TestCase):
 
     def test_register(self):
         app, registry = self._create_app()
-        self._create_register(registry.register)
+        self._create_register(registry.register, app)
         ret = app.test_client().get("/api/v1/resources.json")
         data = json.loads(s(ret.data))
         self.assertEqual(data['swaggerVersion'], '1.3')
@@ -128,3 +167,79 @@ class TestDecorators(unittest.TestCase):
                          ['/cheese.{format}', '/holyHandGrenade.{format}'])
         self.assertEqual(data['apiVersion'], '1.0')
         self.assertEqual(data['models'], {})
+        # test cheese
+        ret = app.test_client().get("/api/v1/cheese.json")
+        data = json.loads(s(ret.data))
+        self.assertEqual(data["resourcePath"], "cheese")
+        self.assertEqual([x.endpoint for x in app.url_map.iter_rules()],
+                         ['holyHandGrenade',
+                          'resources',
+                          'cheese',
+                          'get_a_holy_hand_grenade',
+                          'toss_the_grenade',
+                          'get_cheese',
+                          'static'])
+
+    def test_add_register(self):
+        app, registry = self._create_app()
+        self._create_add_register(registry.add_register, app)
+        ret = app.test_client().get("/api/v1/resources.json")
+        data = json.loads(s(ret.data))
+        self.assertEqual(data['swaggerVersion'], '1.3')
+        self.assertEqual(data['basePath'], 'http://localhost:5000/api/v1')
+        self.assertEqual(sorted([x["path"] for x in data['apis']]),
+                         ['/cheese.{format}'])
+        self.assertEqual(data['apiVersion'], '1.0')
+        self.assertEqual(data['models'], {})
+        # test cheese
+        ret = app.test_client().get("/api/v1/cheese.json")
+        data = json.loads(s(ret.data))
+        self.assertEqual(data["resourcePath"], "cheese")
+
+    def test_register_blueprint(self):
+        app, registry = self._create_app()
+        bp = self._create_blueprint(app)
+        self._create_register(registry.register, app, bp)
+        ret = app.test_client().get("/api/v1/resources.json")
+        data = json.loads(s(ret.data))
+        self.assertEqual(data['swaggerVersion'], '1.3')
+        self.assertEqual(data['basePath'], 'http://localhost:5000/api/v1')
+        self.assertEqual(sorted([x["path"] for x in data['apis']]),
+                         ['/cheese.{format}', '/holyHandGrenade.{format}'])
+        self.assertEqual(data['apiVersion'], '1.0')
+        self.assertEqual(data['models'], {})
+        # test cheese
+        ret = app.test_client().get("/api/v1/cheese.json")
+        data = json.loads(s(ret.data))
+        self.assertEqual(data["resourcePath"], "cheese")
+        # FIXME this is broken urls shouldn't be added multiple times
+        self.assertEqual([x.endpoint for x in app.url_map.iter_rules()],
+                         ['holyHandGrenade',
+                          'resources',
+                          'cheese',
+                          'blueish.get_a_holy_hand_grenade',
+                          'blueish.toss_the_grenade',
+                          'blueish.get_cheese',
+                          'static'])
+
+    def test_add_register_blueprint(self):
+        app, registry = self._create_app()
+        bp = self._create_blueprint(app)
+        self._create_add_register(registry.add_register, app, bp)
+        ret = app.test_client().get("/api/v1/resources.json")
+        data = json.loads(s(ret.data))
+        self.assertEqual(data['swaggerVersion'], '1.3')
+        self.assertEqual(data['basePath'], 'http://localhost:5000/api/v1')
+        self.assertEqual(sorted([x["path"] for x in data['apis']]),
+                         ['/cheese.{format}'])
+        self.assertEqual(data['apiVersion'], '1.0')
+        self.assertEqual(data['models'], {})
+        # test cheese
+        ret = app.test_client().get("/api/v1/cheese.json")
+        data = json.loads(s(ret.data))
+        self.assertEqual(data["resourcePath"], "cheese")
+        self.assertEqual([x.endpoint for x in app.url_map.iter_rules()],
+                         ['resources',
+                          'cheese',
+                          'blueish.get_cheese',
+                          'static'])